refactor: CartItem을 식별 가능한 컴포넌트들로 분리
